rename weatherDB_manager to weatherDB
