+++ HydroApps/urls.py
@@ -1,4 +1,4 @@
-"""weatherDB_manager URL Configuration
+"""weatherDB URL Configuration
 
 The `urlpatterns` list routes URLs to views. For more information please see:
     https://docs.djangoproject.com/en/3.2/topics/http/urls/
@@ -16,7 +16,7 @@ Including another URLconf
 from django.urls import path
 from .views import home_view
 
-#from weatherDB_manager.urls import urlpatterns as weatherDB_manager_urls
+#from weatherDB.urls import urlpatterns as weatherDB_urls
 app_name = 'HydroApps'
 urlpatterns = [
     path("", home_view, name="home")
