+++ klimzuk/urls.py
@@ -1,4 +1,4 @@
-"""weatherDB_manager URL Configuration
+"""weatherDB URL Configuration
 
 The `urlpatterns` list routes URLs to views. For more information please see:
     https://docs.djangoproject.com/en/3.2/topics/http/urls/
