+++ weatherDB/classes/weatherDB/lib/max_fun/import_DWD.py
