+++ weatherDB/static/weatherDB/css/images/markers-matte.png
