+++ weatherDB/static/weatherDB/css/images/markers-matte@2x.png
