+++ weatherDB/static/weatherDB/css/images/markers-plain.png
