+++ weatherDB/static/weatherDB/css/images/markers-shadow.png
