+++ weatherDB/static/weatherDB/css/images/markers-shadow@2x.png
