+++ weatherDB/static/weatherDB/css/images/markers-soft.png
