+++ weatherDB/static/weatherDB/css/images/markers-soft@2x.png
