+++ weatherDB/static/weatherDB/img/marker-icon-blue.png
