+++ weatherDB/static/weatherDB/img/marker-icon-green.png
