+++ weatherDB/static/weatherDB/img/marker-icon-orange.png
