+++ weatherDB/static/weatherDB/img/marker-icon.png
