+++ weatherDB/static/weatherDB/js/get_ts_map.js
