+++ weatherDB/static/weatherDB/js/leaflet/leaflet.awesome-markers.js
