+++ weatherDB/templates/weatherDB/base.html
